feat: 新增admin路由处理

+++ apps/jsonlee-admin/src/pages/scan/Layout/SupperAdmin/Menu/index.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect, useMemo, useRef } from 'react';
-import { Button, Space, Switch, message } from 'antd';
-import EditModal from '@/pages/Menu/EditModal';
+import { Button, message, Modal, Space, Switch } from 'antd';
+import EditModal from '@/pages/scan/Layout/SupperAdmin/Menu/EditModal';
 import { MenuApi } from '@/api/modules/menu';
 import { getMenuData } from '@/store/modules/menu';
 import { useAppDispatch, useAppSelector } from '@/hooks/store';
@@ -24,39 +24,47 @@ const Menu = () => {
         key: 'expandable',
         align: 'center',
         width: 30,
+        fixed: 'left',
         render: (_, _record) => <></>,
       },
       {
         title: 'id',
         dataIndex: 'id',
+        width: 80,
       },
       {
         title: '标题',
         dataIndex: 'title',
         ellipsis: true,
+        width: 120,
         render: (_, record) => <span>{record.meta?.title}</span>,
       },
       {
         title: '图标',
         dataIndex: 'icon',
+        width: 80,
         render: (_, record) => <Icon name={record.meta?.icon}>--</Icon>,
       },
       {
         title: '别名',
         dataIndex: 'name',
         ellipsis: true,
+        width: 100,
       },
       {
         title: '路由地址',
         dataIndex: 'path',
+        width: 120,
       },
       {
         title: '路由文件地址',
         dataIndex: 'component',
+        width: 160,
       },
       {
         title: '是否隐藏',
         dataIndex: 'hidden',
+        width: 90,
         render: (_, record) => (
           <Switch
             checkedChildren="隐藏"
@@ -68,31 +76,48 @@ const Menu = () => {
       {
         title: '排序',
         dataIndex: 'sort',
+        width: 80,
       },
       {
         title: '操作',
+        fixed: 'right',
+        width: 180,
         render: (_, record: MenuItem) => (
           <Space>
-            <Button
-              type={'link'}
+            <a
+              className={'text-[#1677ff]'}
+              onClick={() => {
+                editRef.current?.open('新增菜单', { parentId: record.id });
+              }}
+            >
+              添加子菜单
+            </a>
+            <a
+              className={'text-[#1677ff]'}
               onClick={() => {
                 editRef.current?.open('修改菜单', record);
               }}
             >
               编辑
-            </Button>
-            <Button
-              type={'link'}
-              danger
+            </a>
+            <a
+              className={'text-[#ff4d4f] hover:text-red-300'}
               onClick={() => {
-                menuApi.delete(record.id).then(() => {
-                  message.success('删除成功');
-                  dispatch(getMenuData());
+                Modal.confirm({
+                  title: '删除菜单',
+                  content: '确认删除该菜单吗？',
+                  okType: 'danger',
+                  onOk: () => {
+                    menuApi.delete(record.id).then(() => {
+                      message.success('删除成功');
+                      dispatch(getMenuData());
+                    });
+                  },
                 });
               }}
             >
               删除
-            </Button>
+            </a>
           </Space>
         ),
       },
